+++ crates/rome_formatter/src/js/auxiliary/module.rs
@@ -1,8 +1,8 @@
+use crate::utils::format_interpreter;
 use crate::{
-    format_elements, formatter_traits::FormatTokenAndNode, hard_line_break,
-    interpreter::format_interpreter, FormatElement, FormatResult, Formatter, ToFormatElement,
+    format_elements, formatter_traits::FormatTokenAndNode, hard_line_break, FormatElement,
+    FormatResult, Formatter, ToFormatElement,
 };
-
 use rslint_parser::ast::JsModule;
 
 impl ToFormatElement for JsModule {
