+++ crates/rome_formatter/src/lib.rs
@@ -51,8 +51,7 @@ mod cst;
 mod format_element;
 mod format_elements;
 mod formatter;
-pub mod formatter_traits;
-pub(crate) mod interpreter;
+pub(crate) mod formatter_traits;
 mod intersperse;
 mod js;
 mod printer;
