+++ crates/rome_formatter/src/utils.rs
@@ -1,6 +1,10 @@
 use crate::formatter_traits::FormatOptionalTokenAndNode;
-use crate::{format_elements, space_token, FormatElement, FormatResult, Formatter};
+use crate::{
+    empty_element, format_elements, hard_line_break, space_token, FormatElement, FormatResult,
+    Formatter,
+};
 use rslint_parser::ast::JsInitializerClause;
+use rslint_parser::SyntaxToken;
 
 /// Utility function to format the node [rslint_parser::ast::JsInitializerClause]
 pub(crate) fn format_initializer_clause(
@@ -10,4 +14,15 @@ pub(crate) fn format_initializer_clause(
     initializer.format_with_or_empty(formatter, |initializer| {
         format_elements![space_token(), initializer]
     })
+}
+
+pub(crate) fn format_interpreter(
+    interpreter: Option<SyntaxToken>,
+    formatter: &Formatter,
+) -> FormatResult<FormatElement> {
+    interpreter.format_with_or(
+        formatter,
+        |interpreter| format_elements![interpreter, hard_line_break()],
+        empty_element,
+    )
 }
